requirements.txt
asgiref
autopep8
django==2.2.12
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
mysqlclient==2.0.2
django-select2==5.8.10
django-widget-tweaks==1.4.8
django-import-export==2.4.0
pillow==7.1.2
django-jet==1.0.8
psycopg2